upd: правки в service

+++ src/main/java/ru/yandex/practicum/filmorate/service/filmService/FilmServiceImpl.java
@@ -1,38 +1,58 @@
 package ru.yandex.practicum.filmorate.service.filmService;
 
+import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.storage.filmStorage.FilmStorage;
+import ru.yandex.practicum.filmorate.storage.userStorage.UserStorage;
 
 import java.time.LocalDate;
 import java.util.List;
 
 @Service
 @Slf4j
+@AllArgsConstructor
 public class FilmServiceImpl implements FilmService {
 
     private final FilmStorage filmStorage;
+    private final UserStorage userStorage;
     private static final LocalDate MIN_RELEASE_DATE = LocalDate.of(1895, 12, 28);
-
-    @Autowired
-    public FilmServiceImpl(FilmStorage filmStorage) {
-        this.filmStorage = filmStorage;
-    }
 
     @Override
     public void addLike(Long filmId, Long userId) {
-        log.info("Лайн для фильма" + filmId + "от юзера " + userId + " добавлен");
-        filmStorage.getFilm(filmId).setLikes(userId);
+        Film film = filmStorage.getFilm(filmId);
+        if (film == null) {
+            log.warn("Ошибка: фильм с id={} не найден.", filmId);
+            throw new NotFoundException("Фильм с id: " + filmId + " не найден.");
+        }
+        if (userStorage.getUser(userId) == null) {
+            log.warn("Ошибка: пользователь с id={} не найден.", userId);
+            throw new NotFoundException("Пользователь с id: " + userId + " не найден.");
+        }
+        film.getLikes().add(userId);
+        log.info("Пользователь с id={} добавил лайк к фильму с id={}.", userId, filmId);
     }
 
     @Override
     public void removeLike(Long filmId, Long userId) {
-        log.info("Лайн для фильма" + filmId + "от юзера " + userId + " удален");
-        filmStorage.getFilm(filmId).deleteLike((userId));
+        Film film = filmStorage.getFilm(filmId);
+        if (film == null) {
+            log.warn("Ошибка: фильм с id={} не найден.", filmId);
+            throw new NotFoundException("Фильм с id: " + filmId + " не найден.");
+        }
+        if (userStorage.getUser(userId) == null) {
+            log.warn("Ошибка: пользователь с id={} не найден.", userId);
+            throw new NotFoundException("Пользователь с id: " + userId + " не найден.");
+        }
+        if (!film.getLikes().contains(userId)) {
+            log.warn("Ошибка: лайк от пользователя с id={} не найден.", userId);
+            throw new ValidationException("Пользователь с id: " + userId + " не ставил лайк фильму с id: " + filmId);
+        }
+        film.getLikes().remove(userId);
+        log.info("Пользователь с id={} удалил лайк к фильму с id={}.", userId, filmId);
     }
 
     @Override
@@ -45,31 +65,41 @@ public class FilmServiceImpl implements FilmService {
 
     @Override
     public void addingFilm(Film film) {
+        validateFilm(film);
         log.info("Фильм добавлен");
         filmStorage.addingFilm(film);
     }
 
     @Override
     public void updateFilm(Film film) {
-        if (!filmStorage.getAllFilm().contains(film.getId())) {
+        if (filmStorage.getFilm((long) film.getId()) == null) {
             log.warn("Фильм не найден");
             throw new NotFoundException("Фильм не найден");
         }
-        log.info("Фильм " + film + " добавлен.");
+        validateFilm(film);
+        log.info("Фильм с id={} обновлен.", film.getId());
         filmStorage.updateFilm(film);
-
     }
 
     @Override
     public void deleteFilm(Long filmId) {
-        log.info("Фильм " + filmId + " удалён.");
+        if (filmStorage.getFilm(filmId) == null) {
+            log.warn("Ошибка: фильм с id={} не найден.", filmId);
+            throw new NotFoundException("Фильм с id: " + filmId + " не найден.");
+        }
+        log.info("Фильм с id={} удалён.", filmId);
         filmStorage.deleteFilm(filmId);
     }
 
     @Override
     public Film getFilm(Long filmId) {
-        log.info("Фильм " + filmId + " получен.");
-        return filmStorage.getFilm(filmId);
+        Film film = filmStorage.getFilm(filmId);
+        if (film == null) {
+            log.warn("Фильм с id={} не найден.", filmId);
+            throw new NotFoundException("Фильм с id: " + filmId + " не найден.");
+        }
+        log.info("Фильм с id={} получен.", filmId);
+        return film;
     }
 
     @Override
