+++ src/main/java/ru/yandex/practicum/filmorate/service/userService/UserServiceImpl.java
@@ -4,6 +4,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.userStorage.UserStorage;
 
@@ -25,9 +26,27 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void addFriend(Long userId, Long friendId) {
-        log.info("Друг " + friendId + " пользователя " + userId + "добавлен.");
-        userStorage.getUser(userId).setFriend(friendId);
-        userStorage.getUser(friendId).setFriend(userId);
+        User user = userStorage.getUser(userId);
+        User friend = userStorage.getUser(friendId);
+
+        if (user == null || friend == null) {
+            log.warn("Пользователь не найден");
+            throw new NotFoundException("Один из пользователей не найден");
+        }
+
+        if (user.getFriends().contains(friendId)) {
+            log.warn("Пользователь с id={} уже является другом пользователя с id={}", friendId, userId);
+            throw new ValidationException("Пользователь уже в друзьях");
+        }
+
+        if (userId.equals(friendId)) {
+            log.warn("Ошибка: пользователь не может добавить сам себя в друзья.");
+            throw new ValidationException("Пользователь не может добавить сам себя в друзья.");
+        }
+
+        log.info("Друг {} добавлен пользователю {}.", friendId, userId);
+        user.setFriend(friendId);
+        friend.setFriend(userId);
     }
 
     @Override
@@ -52,7 +71,11 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void addingUser(User user) {
-        log.info("Пользователь " + user + " добавлен.");
+        if (userStorage.getUser((long) user.getId()) != null) {
+            log.warn("Такой пользователь уже есть.");
+            throw new ValidationException("Такой пользователь уже существует.");
+        }
+        log.info("Пользователь {} добавлен.", user);
         userStorage.addingUser(user);
     }
 
@@ -62,19 +85,29 @@ public class UserServiceImpl implements UserService {
             log.warn("Пользователь не найден");
             throw new NotFoundException("Пользователь не найден");
         }
+        log.info("Пользователь {} обновлен.", user);
         userStorage.updateUser(user);
     }
 
     @Override
     public void deleteUser(User user) {
-        log.info("Пользователь " + user + " удалён.");
+        if (userStorage.getUser((long) user.getId()) == null) {
+            log.warn("Пользователь не найден");
+            throw new NotFoundException("Пользователь не найден");
+        }
+        log.info("Пользователь {} удалён.", user);
         userStorage.deleteUser(user);
     }
 
     @Override
     public User getUser(Long userId) {
-        log.info("Пользователь " + userId + " возвращен.");
-        return userStorage.getUser(userId);
+        User user = userStorage.getUser(userId);
+        if (user == null) {
+            log.warn("Пользователь с id={} не найден", userId);
+            throw new NotFoundException("Пользователь с id=" + userId + " не найден");
+        }
+        log.info("Пользователь {} возвращен.", userId);
+        return user;
     }
 
     @Override
@@ -84,21 +117,25 @@ public class UserServiceImpl implements UserService {
     }
 
     public List<Long> findCommonFriends(Long id, Long friendId) {
-        User user = userStorage.getUser(id);
-        User friendUser = userStorage.getUser(friendId);
+        User user = getUser(id);
+        User friendUser = getUser(friendId);
 
         Set<Long> userFriends = new HashSet<>(user.getFriends());
         Set<Long> friendFriends = new HashSet<>(friendUser.getFriends());
 
         userFriends.retainAll(friendFriends);
 
-        log.info("Список друга друзей возвращен");
+        log.info("Список общих друзей между пользователями с id={} и id={} возвращен.", id, friendId);
         return new ArrayList<>(userFriends);
     }
 
     public void validateUser(User user) {
         if (user.getName() == null || user.getName().isBlank()) {
             user.setName(user.getLogin());
+        }
+        if (user.getLogin().contains(" ")) {
+            log.warn("Ошибка в написании логина.");
+            throw new ValidationException("Логин не должен содержать пробелы.");
         }
     }
 }
